+++ event-buffer/event-buffer-core/src/main/java/uk/gov/justice/services/event/buffer/core/repository/subscription/StreamUpdateContext.java
@@ -1,0 +1,11 @@
+package uk.gov.justice.services.event.buffer.core.repository.subscription;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public record StreamUpdateContext(
+        long incomingEventPosition,
+        long currentStreamPosition,
+        long latestKnownStreamPosition,
+        Optional<UUID> streamErrorId) {
+}